Do not use absURL to include scripts
When using this method, the Hugo compiler resolve the host to its default
base URL which is `localhost:<port>`. I don't want to be setting a base
URL of developers.strava.com at this time because it would prevent people
from running the Docker image locally. Instead, we can use relative URLs
to resolve the assets.