Fixes the path of numerous files in the Swagger spec
I noticed yesterday that the Java code generated by Swagger didn't compile. When I reorganized the models a few weeks back, I had to fix the paths of the json files in numerous places. It looks like I haven't been thorough enough in doing that and that the generated code could no longer compile. This change:

- Addresses all the instances of broken references to model files, across the entire API spec
- Adds a test to enforce that the Java code at the very least compiles.

Down the line, we should probably have test suites to validate that the generated code is working across several languages.