Introduces a build step to generate the Swagger specification
Currently, the entire public Swagger specification uses relatives paths (`./`) in JSON references (`$ref`). This is an easy way to get things to work both locally and in production, where the reference is resolved against the path of the top-level specification:

- locally, `./fault.json#Fault` called from `./static/swagger/swagger.json` will resolve to `./static/swagger/fault.json#Fault`
- riIn production, `./fault.json#Fault` called from `https://developers.strava.com/swagger/swagger.json` will resolve to `https://developers.strava.com/swagger/fault.json#Fault`

However, this method prevents other specifications from refering to shared models that exist in production. For instance, `activity.json` embeds a reference to `./photo.json` - note the relative path. A localSwagger specification that refers to `https://developers.strava.com/swagger/activity.json` will in turn try to load `./photo.json` locally and fail to do so. It's unclear to me whether this is as specified in the JSON Ref specification or a bug in Swagger Codegen's implementation but this is what is observed.

In order to allow reuse of shared production models, we should make sure that they embed absolute references to their dependencies. However, we also want to retain the ability to generate documentation or code that contains local modifications and therefore do not refer to the production world at all. This change introduces a build step to generate the Swagger specification and models in either production or local mode. This is a simple mustache-based step with currently a single variable: `reference_prefix` is `.` in local mode and `https://developers.strava.com` in production.

By default, the Dockerfile generates a production build. In specs (which are run before any change can be merged), we use specification generated in local mode to ake sure we accoun for all local changes.