Gets the playground to work
While testing the playgroud, I realized it was incorrectly configured. Since we rely on OAuth2 to authenticate athletes, the playground needs to have a page to handle the redirect that happens after the user authenticates - we had none. On top of that, we need to configure the correct `oauth2RedirectUrl` parameters in the playground configuration, and also amend the API specification to require some level of authentication for every request.