[CP-1770] Add a changelog
Background:
There is currently no changelog in [developers.strava.com](https://developers.strava.com). However, changelogs are important becaduse

Solution:
Add a changelog. A link to the old changelog was already linked on the [developers.strava.com](https://developers.strava.com) homepage, so all that was needed was to create a new markdown file and update the link to be a local reference.

I wanted to make it simple for developers to add changelog entries while ensuring that the content was presented logically and easily readable, so I relied heavily on built-in Markdown conventions and looked around at other well-documented external APIs for stylistic inspiration; the current implementation is similar to that of the [Stripe API](https://stripe.com/docs/upgrades#api-changelog), where dates are bolded and individual entries are separated by unordered lists.

Result:
[developers.strava.com](https://developers.strava.com) has a changelog that is easy to update and easy to understand.

Completes [CP-1770](https://strava.atlassian.net/browse/CP-1770)